+++ src/main/java/org/snipsniper/configwindow/tabs/EditorTab.java
@@ -6,9 +6,7 @@ import org.snipsniper.config.Config;
 import org.snipsniper.config.ConfigHelper;
 import org.snipsniper.configwindow.ConfigWindow;
 import org.snipsniper.configwindow.StampJPanel;
-import org.snipsniper.sceditor.stamps.CubeStamp;
-import org.snipsniper.sceditor.stamps.IStamp;
-import org.snipsniper.sceditor.stamps.StampUtils;
+import org.snipsniper.sceditor.stamps.*;
 import org.snipsniper.utils.Function;
 import org.snipsniper.utils.InfoButton;
 import org.snipsniper.utils.enums.ConfigSaveButtonState;
@@ -95,7 +93,7 @@ public class EditorTab extends JPanel implements ITab{
             if (e.getStateChange() == ItemEvent.SELECTED) {
                 IStamp newStamp = StampUtils.getNewIStampByIndex(stampDropdown.getSelectedIndex(), config, null);
                 row3_stampPreview.setStamp(newStamp);
-                configWindow.setupStampConfigPanel(row3_stampConfig, newStamp, row3_stampPreview, config, onUpdate[0]);
+                setupStampConfigPanel(row3_stampConfig, newStamp, row3_stampPreview, config, onUpdate[0]);
                 saveButtonUpdate[0].run();
             }
         });
@@ -127,12 +125,170 @@ public class EditorTab extends JPanel implements ITab{
                 return saveButtonUpdate[0].run(ConfigSaveButtonState.UPDATE_CLEAN_STATE);
             }
         };
-        configWindow.setupStampConfigPanel(row3_stampConfig, stamp, row3_stampPreview, config, onUpdate[0]);
+        setupStampConfigPanel(row3_stampConfig, stamp, row3_stampPreview, config, onUpdate[0]);
 
 
         add(options);
 
         if(disablePage)
             configWindow.setEnabledAll(options, false, dropdown);
+    }
+
+    private JSpinner setupStampConfigPanelSpinner(Enum configKey, double min, double max, double stepSize, StampJPanel previewPanel, Config config, int stampIndex, Function onUpdate) {
+        JSpinner spinner = new JSpinner(new SpinnerNumberModel(Double.parseDouble(config.getFloat(configKey)+""), min, max, stepSize));
+        spinner.addChangeListener(e -> {
+            config.set(configKey, (int)Double.parseDouble(spinner.getValue().toString()));
+            previewPanel.setStamp(StampUtils.getNewIStampByIndex(stampIndex, config, null));
+            onUpdate.run();
+        });
+        return spinner;
+    }
+
+    private void setupStampConfigPanelSpinnerWithLabel(JPanel panel, String title, Enum configKey, double min, double max, double stepSize, StampJPanel previewPanel, Config config, int stampIndex, GridBagConstraints constraints, String infoText, Function onUpdate) {
+        constraints.gridx = 0;
+        panel.add(configWindow.createJLabel(title, JLabel.RIGHT, JLabel.CENTER), constraints);
+        constraints.gridx = 1;
+        panel.add(setupStampConfigPanelSpinner(configKey, min, max, stepSize,previewPanel, config, stampIndex, onUpdate), constraints);
+        constraints.gridx = 2;
+        panel.add(new InfoButton(infoText), constraints);
+        constraints.gridx = 0;
+    }
+
+    public void setupStampConfigPanel(JPanel panel, IStamp stamp, StampJPanel previewPanel, Config config, Function onUpdate) {
+        panel.removeAll();
+        GridBagConstraints gbc = new GridBagConstraints();
+        gbc.fill = GridBagConstraints.BOTH;
+        gbc.insets = new Insets(0, 4, 0, 4);
+
+        if(stamp instanceof CubeStamp) {
+            panel.add(configWindow.createJLabel("Start color", JLabel.RIGHT, JLabel.CENTER), gbc);
+            gbc.gridx = 1;
+            panel.add(configWindow.setupColorButton("Color", config, ConfigHelper.PROFILE.editorStampCubeDefaultColor, e -> {
+                previewPanel.setStamp(new CubeStamp(config, null));
+                onUpdate.run();
+            }), gbc);
+            gbc.gridx = 2;
+            panel.add(new InfoButton(null), gbc);
+            gbc.gridx = 0;
+
+            panel.add(configWindow.createJLabel("Smart Pixel", JLabel.RIGHT, JLabel.CENTER), gbc);
+            gbc.gridx = 1;
+            JCheckBox smartPixelCheckBox = new JCheckBox();
+            smartPixelCheckBox.setSelected(config.getBool(ConfigHelper.PROFILE.editorStampCubeSmartPixel));
+            smartPixelCheckBox.addActionListener(e -> {
+                config.set(ConfigHelper.PROFILE.editorStampCubeSmartPixel, smartPixelCheckBox.isSelected() + "");
+                onUpdate.run();
+            });
+            panel.add(smartPixelCheckBox, gbc);
+            gbc.gridx = 2;
+            panel.add(new InfoButton(null), gbc);
+            gbc.gridx = 0;
+
+            setupStampConfigPanelSpinnerWithLabel(panel, "Start width", ConfigHelper.PROFILE.editorStampCubeWidth, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CUBE, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Start height", ConfigHelper.PROFILE.editorStampCubeHeight, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CUBE, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Width change speed", ConfigHelper.PROFILE.editorStampCubeWidthSpeed, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CUBE, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Height change speed", ConfigHelper.PROFILE.editorStampCubeHeightSpeed, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CUBE, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Minimum width", ConfigHelper.PROFILE.editorStampCubeWidthMinimum, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CUBE, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Minimum height", ConfigHelper.PROFILE.editorStampCubeHeightMinimum, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CUBE, gbc, null, onUpdate);
+        } else if(stamp instanceof CounterStamp) {
+            panel.add(configWindow.createJLabel("Start color", JLabel.RIGHT, JLabel.CENTER), gbc);
+            gbc.gridx = 1;
+            panel.add(configWindow.setupColorButton("Color", config, ConfigHelper.PROFILE.editorStampCounterDefaultColor, e -> previewPanel.setStamp(new CounterStamp(config))), gbc);
+            gbc.gridx = 2;
+            panel.add(new InfoButton(null), gbc);
+            gbc.gridx = 0;
+
+            setupStampConfigPanelSpinnerWithLabel(panel, "Start width", ConfigHelper.PROFILE.editorStampCounterWidth, 1, 999, 1, previewPanel, config, StampUtils.INDEX_COUNTER, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Start height", ConfigHelper.PROFILE.editorStampCounterHeight, 1, 999, 1, previewPanel, config, StampUtils.INDEX_COUNTER, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "General change speed", ConfigHelper.PROFILE.editorStampCounterSpeed, 1, 999, 1, previewPanel, config, StampUtils.INDEX_COUNTER, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Width change speed", ConfigHelper.PROFILE.editorStampCounterWidthSpeed, 1, 999, 1, previewPanel, config, StampUtils.INDEX_COUNTER, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Height change speed", ConfigHelper.PROFILE.editorStampCounterHeightSpeed, 1, 999, 1, previewPanel, config, StampUtils.INDEX_COUNTER, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Minimum width", ConfigHelper.PROFILE.editorStampCounterWidthMinimum, 1, 999, 1, previewPanel, config, StampUtils.INDEX_COUNTER, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Minimum height", ConfigHelper.PROFILE.editorStampCounterHeightMinimum, 1, 999, 1, previewPanel, config, StampUtils.INDEX_COUNTER, gbc, null, onUpdate);
+
+            panel.add(configWindow.createJLabel("Solid color", JLabel.RIGHT, JLabel.CENTER), gbc);
+            JCheckBox cbSolidColor = new JCheckBox();
+            cbSolidColor.setSelected(config.getBool(ConfigHelper.PROFILE.editorStampCounterSolidColor));
+            cbSolidColor.addChangeListener(e -> {
+                config.set(ConfigHelper.PROFILE.editorStampCounterSolidColor, cbSolidColor.isSelected() + "");
+                previewPanel.setStamp(new CounterStamp(config));
+                onUpdate.run();
+            });
+            gbc.gridx = 1;
+            panel.add(cbSolidColor, gbc);
+            gbc.gridx = 2;
+            panel.add(new InfoButton(null), gbc);
+
+            gbc.gridx = 0;
+            panel.add(configWindow.createJLabel("Stamp Border", JLabel.RIGHT, JLabel.CENTER), gbc);
+            gbc.gridx = 1;
+            JCheckBox cbBorder = new JCheckBox();
+            cbBorder.setSelected(config.getBool(ConfigHelper.PROFILE.editorStampCounterBorderEnabled));
+            cbBorder.addChangeListener(e -> {
+                config.set(ConfigHelper.PROFILE.editorStampCounterBorderEnabled, cbBorder.isSelected() + "");
+                previewPanel.setStamp(new CounterStamp(config));
+                onUpdate.run();
+            });
+            panel.add(cbBorder, gbc);
+            gbc.gridx = 2;
+            panel.add(new InfoButton(null), gbc);
+
+            setupStampConfigPanelSpinnerWithLabel(panel, "Font size modifier", ConfigHelper.PROFILE.editorStampCounterFontSizeModifier, 0.1, 10, 0.01D, previewPanel, config, StampUtils.INDEX_COUNTER, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Border modifier", ConfigHelper.PROFILE.editorStampCounterBorderModifier, 1, 999, 1, previewPanel, config, StampUtils.INDEX_COUNTER, gbc, null, onUpdate);
+        } else if(stamp instanceof CircleStamp) {
+            panel.add(configWindow.createJLabel("Start color", JLabel.RIGHT, JLabel.CENTER), gbc);
+            gbc.gridx = 1;
+            panel.add(configWindow.setupColorButton("Color", config, ConfigHelper.PROFILE.editorStampCircleDefaultColor, e -> previewPanel.setStamp(new CircleStamp(config))), gbc);
+            gbc.gridx = 2;
+            panel.add(new InfoButton(null), gbc);
+
+            setupStampConfigPanelSpinnerWithLabel(panel, "Start width", ConfigHelper.PROFILE.editorStampCircleWidth, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CIRCLE, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Start height", ConfigHelper.PROFILE.editorStampCircleHeight, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CIRCLE, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "General change speed", ConfigHelper.PROFILE.editorStampCircleSpeed, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CIRCLE, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Width change speed", ConfigHelper.PROFILE.editorStampCircleWidthSpeed, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CIRCLE, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Height change speed", ConfigHelper.PROFILE.editorStampCircleHeightSpeed, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CIRCLE, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Minimum width", ConfigHelper.PROFILE.editorStampCircleWidthMinimum, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CIRCLE, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Minimum height", ConfigHelper.PROFILE.editorStampCircleHeightMinimum, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CIRCLE, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Thickness", ConfigHelper.PROFILE.editorStampCircleThickness, 1, 999, 1, previewPanel, config, StampUtils.INDEX_CIRCLE, gbc, null, onUpdate);
+        } else if(stamp instanceof SimpleBrush) {
+            panel.add(configWindow.createJLabel("Start color", JLabel.RIGHT, JLabel.CENTER), gbc);
+            gbc.gridx = 1;
+            panel.add(configWindow.setupColorButton("Color", config, ConfigHelper.PROFILE.editorStampSimpleBrushDefaultColor, e -> previewPanel.setStamp(new SimpleBrush(config, null))), gbc);
+            gbc.gridx = 2;
+            panel.add(new InfoButton("text"), gbc);
+
+            setupStampConfigPanelSpinnerWithLabel(panel, "Brush size", ConfigHelper.PROFILE.editorStampSimpleBrushSize, 1, 999, 1, previewPanel, config, StampUtils.INDEX_SIMPLE_BRUSH, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Brush size change speed", ConfigHelper.PROFILE.editorStampSimpleBrushSizeSpeed, 1, 999, 1, previewPanel, config, StampUtils.INDEX_SIMPLE_BRUSH, gbc, null, onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Line point distance", ConfigHelper.PROFILE.editorStampSimpleBrushDistance, 1, 999, 1, previewPanel, config, StampUtils.INDEX_SIMPLE_BRUSH, gbc, null, onUpdate);
+            panel.add(new JPanel()); //Padding
+        } else if(stamp instanceof TextStamp) {
+            panel.add(configWindow.createJLabel("Start color", JLabel.RIGHT, JLabel.CENTER), gbc);
+            gbc.gridx = 1;
+            panel.add(configWindow.setupColorButton("Color", config, ConfigHelper.PROFILE.editorStampTextDefaultColor, e -> previewPanel.setStamp(new TextStamp(config, null))), gbc);
+            gbc.gridx = 2;
+            panel.add(new InfoButton("text"), gbc);
+
+            setupStampConfigPanelSpinnerWithLabel(panel, "Default font size", ConfigHelper.PROFILE.editorStampTextDefaultFontSize, 1, 999, 1, previewPanel, config, StampUtils.INDEX_TEXT, gbc, "text", onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Font size change speed", ConfigHelper.PROFILE.editorStampTextDefaultSpeed, 1, 999, 1, previewPanel, config, StampUtils.INDEX_TEXT, gbc, "text", onUpdate);
+            for(int i = 0; i < 6; i++) panel.add(new JPanel(), gbc); //Padding
+            //TODO: Draw it in the middle, possibly by giving TextStamp a getTextWidth() function and adding an edgecase to the Stamp Renderer, to move it to the left
+        } else if(stamp instanceof RectangleStamp) {
+            panel.add(configWindow.createJLabel("Start color", JLabel.RIGHT, JLabel.CENTER), gbc);
+            gbc.gridx = 1;
+            panel.add(configWindow.setupColorButton("Color", config, ConfigHelper.PROFILE.editorStampRectangleDefaultColor, e -> previewPanel.setStamp(new RectangleStamp(config))), gbc);
+            gbc.gridx = 2;
+            panel.add(new InfoButton("text"), gbc);
+
+            setupStampConfigPanelSpinnerWithLabel(panel, "Start width", ConfigHelper.PROFILE.editorStampRectangleWidth, 1, 999, 1, previewPanel, config, StampUtils.INDEX_RECTANGLE, gbc, "text", onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Start height", ConfigHelper.PROFILE.editorStampRectangleHeight, 1, 999, 1, previewPanel, config, StampUtils.INDEX_RECTANGLE, gbc, "text", onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Width change speed", ConfigHelper.PROFILE.editorStampRectangleWidthSpeed, 1, 999, 1, previewPanel, config, StampUtils.INDEX_RECTANGLE, gbc, "text", onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Height change speed", ConfigHelper.PROFILE.editorStampRectangleHeightSpeed, 1, 999, 1, previewPanel, config, StampUtils.INDEX_RECTANGLE, gbc, "text", onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Minimum width", ConfigHelper.PROFILE.editorStampRectangleWidthMinimum, 1, 999, 1, previewPanel, config, StampUtils.INDEX_RECTANGLE, gbc, "text", onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Minimum height", ConfigHelper.PROFILE.editorStampRectangleHeightMinimum, 1, 999, 1, previewPanel, config, StampUtils.INDEX_RECTANGLE, gbc, "text", onUpdate);
+            setupStampConfigPanelSpinnerWithLabel(panel, "Thickness", ConfigHelper.PROFILE.editorStampRectangleThickness, 1, 999, 1, previewPanel, config, StampUtils.INDEX_RECTANGLE, gbc, "text", onUpdate);
+        } else {
+            panel.add(configWindow.createJLabel("Coming soon", JLabel.CENTER, JLabel.CENTER));
+            for (int i = 0; i < 15; i++) panel.add(new JLabel());
+        }
     }
 }
