+++ src/main/java/org/snipsniper/configwindow/tabs/GlobalTab.java
@@ -26,7 +26,7 @@ import java.util.zip.ZipInputStream;
 import java.util.zip.ZipOutputStream;
 
 public class GlobalTab extends JPanel implements ITab{
-    private ConfigWindow configWindow;
+    private final ConfigWindow configWindow;
 
     public GlobalTab(ConfigWindow configWindow) {
         this.configWindow = configWindow;
